(MINOR) Refactor instrument ingestion and relocate currency supply logic (#902)
