+++ src/main/java/com/verlumen/tradestream/ingestion/BUILD
@@ -16,15 +16,11 @@ java_binary(
         ":real_time_data_ingestion",
         ":real_time_data_ingestion_impl",
         "//src/main/java/com/verlumen/tradestream/execution:run_mode",
+        "//src/main/java/com/verlumen/tradestream/instruments:instruments_module",
         "//third_party:argparse4j",
         "//third_party:flogger",
         "//third_party:guice",
     ],
-)
-
-java_library(
-    name = "coin_market_cap_config",
-    srcs = ["CoinMarketCapConfig.java"],
 )
 
 container_structure_test(
@@ -32,39 +28,6 @@ container_structure_test(
     configs = ["container-structure-test.yaml"],
     image = ":image",
     tags = ["requires-docker"],
-)
-
-java_library(
-    name = "currency_pair_metadata",
-    srcs = ["CurrencyPairMetadata.java"],
-    deps = [
-        "//src/main/java/com/verlumen/tradestream/instruments:currency_pair",
-    ],
-)
-
-java_library(
-    name = "currency_pair_supply",
-    srcs = ["CurrencyPairSupply.java"],
-    deps = [
-        ":currency_pair_metadata",
-        "//src/main/java/com/verlumen/tradestream/instruments:currency_pair",
-        "//third_party:guava",
-        "//third_party:auto_value",
-    ],
-)
-
-java_library(
-    name = "currency_pair_supply_provider",
-    srcs = ["CurrencyPairSupplyProvider.java"],
-    deps = [
-        ":coin_market_cap_config",
-        ":currency_pair_metadata",
-        ":currency_pair_supply",
-        "//src/main/java/com/verlumen/tradestream/http:http_client",
-        "//third_party:gson",
-        "//third_party:guava",
-        "//third_party:guice",
-    ],
 )
 
 genrule(
@@ -114,13 +77,13 @@ java_library(
     name = "ingestion_module",
     srcs = ["IngestionModule.java"],
     deps = [
-        ":coin_market_cap_config",
-        ":currency_pair_supply",
-        ":currency_pair_supply_provider",
         ":ingestion_config",
         ":real_time_data_ingestion",
         ":real_time_data_ingestion_impl",
         "//src/main/java/com/verlumen/tradestream/execution:run_mode",
+        "//src/main/java/com/verlumen/tradestream/instruments:coin_market_cap_config",
+        "//src/main/java/com/verlumen/tradestream/instruments:currency_pair_supply",
+        "//src/main/java/com/verlumen/tradestream/instruments:currency_pair_supply_provider",
         "//src/main/java/com/verlumen/tradestream/http:http_module",
         "//src/main/java/com/verlumen/tradestream/kafka:kafka_module",
         "//src/main/java/com/verlumen/tradestream/marketdata:market_data_module",
@@ -151,10 +114,10 @@ java_library(
     name = "real_time_data_ingestion_impl",
     srcs = ["RealTimeDataIngestionImpl.java"],
     deps = [
-        ":currency_pair_supply",
         ":real_time_data_ingestion",
         "//protos:marketdata_java_proto",
         "//src/main/java/com/verlumen/tradestream/instruments:currency_pair",
+        "//src/main/java/com/verlumen/tradestream/instruments:currency_pair_supply",
         "//src/main/java/com/verlumen/tradestream/marketdata:exchange_streaming_client",
         "//src/main/java/com/verlumen/tradestream/marketdata:trade_publisher",
         "//third_party:flogger",
