--- src/main/java/com/verlumen/tradestream/ingestion/CurrencyPairSupply.java
@@ -1,23 +1,0 @@
-package com.verlumen.tradestream.ingestion;
-
-import static com.google.common.collect.ImmutableList.toImmutableList;
-
-import com.google.auto.value.AutoValue;
-import com.google.common.collect.ImmutableList;
-import com.verlumen.tradestream.instruments.CurrencyPair;
-
-@AutoValue
-abstract class CurrencyPairSupply {
-  static CurrencyPairSupply create(ImmutableList<CurrencyPairMetadata> metadataList) {
-    return new AutoValue_CurrencyPairSupply(metadataList);
-  }
-
-  abstract ImmutableList<CurrencyPairMetadata> metadataList();
-
-  ImmutableList<CurrencyPair> currencyPairs() {
-    return metadataList()
-      .stream()
-      .map(CurrencyPairMetadata::currencyPair)
-      .collect(toImmutableList());
-  }
-}
