+++ src/main/java/com/verlumen/tradestream/ingestion/IngestionConfig.java
@@ -3,8 +3,6 @@ package com.verlumen.tradestream.ingestion;
 import com.verlumen.tradestream.execution.RunMode;
 
 record IngestionConfig(
-    String coinMarketCapApiKey,
-    int topCryptocurrencyCount,
     String exchangeName,
     RunMode runMode,
     String kafkaBootstrapServers,
