+++ src/main/java/com/verlumen/tradestream/ingestion/RealTimeDataIngestionImpl.java
@@ -11,6 +11,7 @@ import com.google.common.flogger.FluentLogger;
 import com.google.inject.Inject;
 import com.google.inject.Provider;
 import com.verlumen.tradestream.instruments.CurrencyPair;
+import com.verlumen.tradestream.instruments.CurrencyPairSupply;
 import com.verlumen.tradestream.marketdata.Trade;
 import com.verlumen.tradestream.marketdata.ExchangeStreamingClient;
 import com.verlumen.tradestream.marketdata.TradePublisher;
