+++ src/main/java/com/verlumen/tradestream/instruments/CoinMarketCapConfig.java
@@ -1,4 +1,4 @@
-package com.verlumen.tradestream.ingestion;
+package com.verlumen.tradestream.instruments;
 
 record CoinMarketCapConfig(int topN, String apiKey) {
   static CoinMarketCapConfig create(int topN, String apiKey) {
