+++ src/main/java/com/verlumen/tradestream/instruments/CurrencyPairSupply.java
@@ -1,0 +1,10 @@
+package com.verlumen.tradestream.instruments;
+
+import com.google.common.collect.ImmutableList;
+import java.util.function.Supplier;
+
+public interface CurrencyPairSupply extends Supplier<ImmutableList<CurrencyPair>> {
+  default ImmutableList<CurrencyPair> currencyPairs() {
+    return get();
+  }
+}
